1. implemented wallet service for create and get wallet. 2. implemented create chapter access, get chapter access. 3. implemented purchase chapter

+++ paymentservice/src/main/java/com/ncs/nusiss/paymentservice/entity/ChapterPointsTransfer.java
@@ -20,5 +20,5 @@ public class ChapterPointsTransfer extends Auditable{
     @NotNull
     private String toWalletId;
     @NotNull
-    private Double pointsTransferred;
+    private Integer pointsTransferred;
 }
